+++ Unosquare.CodeChallenge/WarehouseManagement/src/app/components/companies/companies.component.ts
@@ -4,7 +4,7 @@ import { BehaviorSubject, Subscriber, Subscription } from 'rxjs';
 import { ICompany } from '../../interfaces/ICompany';
 import { Company } from '../../models/company.model';
 import { CompaniesService } from '../../services/companies.service';
-import { AddcompanyComponent } from './parts/addcompany/addcompany.component';
+import { AddcompanyComponent } from "./parts/addcompany/AddcompanyComponent";
 
 @Component({
   selector: 'app-companies',
